# Conflicts:
#	common/src/main/java/mtr/client/CustomResources.java
#	common/src/main/java/mtr/client/TrainClientRegistry.java
#	common/src/main/java/mtr/data/Train.java
#	common/src/main/java/mtr/data/TrainClient.java
#	common/src/main/java/mtr/data/TrainType.java
#	common/src/main/java/mtr/model/ModelR211.java
#	common/src/main/java/mtr/model/ModelTrainBase.java
#	common/src/main/java/mtr/render/RenderTrains.java

+++ common/src/main/java/mtr/item/ItemNodeModifierBase.java
@@ -5,13 +5,12 @@ import mtr.block.BlockNode;
 import mtr.data.RailAngle;
 import mtr.data.RailwayData;
 import mtr.data.TransportMode;
+import mtr.mappings.Text;
 import net.minecraft.ChatFormatting;
 import net.minecraft.core.BlockPos;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.Style;
-import net.minecraft.network.chat.TranslatableComponent;
-import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
@@ -23,7 +22,7 @@ import net.minecraft.world.level.block.state.BlockState;
 
 import java.util.List;
 
-public abstract class ItemNodeModifierBase extends Item {
+public abstract class ItemNodeModifierBase extends ItemBlockClickingBase {
 
 	public final boolean forNonContinuousMovementNode;
 	public final boolean forContinuousMovementNode;
@@ -40,63 +39,55 @@ public abstract class ItemNodeModifierBase extends Item {
 	}
 
 	@Override
-	public InteractionResult useOn(UseOnContext context) {
-		final Level world = context.getLevel();
-		if (!world.isClientSide) {
-			final RailwayData railwayData = RailwayData.getInstance(world);
-			final BlockPos posStart = context.getClickedPos();
-			final BlockState stateStart = world.getBlockState(posStart);
-			final Block blockStart = stateStart.getBlock();
-
-			if (railwayData != null && blockStart instanceof BlockNode && (((BlockNode) blockStart).transportMode.continuousMovement && forContinuousMovementNode || !((BlockNode) blockStart).transportMode.continuousMovement && forNonContinuousMovementNode)) {
-				final CompoundTag compoundTag = context.getItemInHand().getOrCreateTag();
-
-				if (compoundTag.contains(TAG_POS) && compoundTag.contains(TAG_TRANSPORT_MODE)) {
-					final BlockPos posEnd = BlockPos.of(compoundTag.getLong(TAG_POS));
-					final BlockState stateEnd = world.getBlockState(posEnd);
-
-					if (stateEnd.getBlock() instanceof BlockNode && ((BlockNode) blockStart).transportMode.toString().equals(compoundTag.getString(TAG_TRANSPORT_MODE))) {
-						final Player player = context.getPlayer();
-
-						if (isConnector) {
-							if (!posStart.equals(posEnd)) {
-								final float angle1 = BlockNode.getAngle(stateStart);
-								final float angle2 = BlockNode.getAngle(stateEnd);
-
-								final float angleDifference = (float) Math.toDegrees(Math.atan2(posEnd.getZ() - posStart.getZ(), posEnd.getX() - posStart.getX()));
-								final RailAngle railAngleStart = RailAngle.fromAngle(angle1 + (RailAngle.similarFacing(angleDifference, angle1) ? 0 : 180));
-								final RailAngle railAngleEnd = RailAngle.fromAngle(angle2 + (RailAngle.similarFacing(angleDifference, angle2) ? 180 : 0));
-
-								onConnect(world, context.getItemInHand(), ((BlockNode) blockStart).transportMode, stateStart, stateEnd, posStart, posEnd, railAngleStart, railAngleEnd, player, railwayData);
-							}
-						} else {
-							onRemove(world, posStart, posEnd, railwayData);
-						}
-					}
-
-					compoundTag.remove(TAG_POS);
-					compoundTag.remove(TAG_TRANSPORT_MODE);
-				} else {
-					compoundTag.putLong(TAG_POS, posStart.asLong());
-					compoundTag.putString(TAG_TRANSPORT_MODE, ((BlockNode) blockStart).transportMode.toString());
-				}
-
-				return InteractionResult.SUCCESS;
-			} else {
-				return InteractionResult.FAIL;
-			}
-		} else {
-			return super.useOn(context);
-		}
-	}
-
-	@Override
 	public void appendHoverText(ItemStack stack, Level level, List<Component> tooltip, TooltipFlag tooltipFlag) {
 		final CompoundTag compoundTag = stack.getOrCreateTag();
 		final long posLong = compoundTag.getLong(TAG_POS);
 		if (posLong != 0) {
-			tooltip.add(new TranslatableComponent("tooltip.mtr.selected_block", BlockPos.of(posLong).toShortString()).setStyle(Style.EMPTY.withColor(ChatFormatting.GOLD)));
+			tooltip.add(Text.translatable("tooltip.mtr.selected_block", BlockPos.of(posLong).toShortString()).setStyle(Style.EMPTY.withColor(ChatFormatting.GOLD)));
 		}
+	}
+
+	@Override
+	protected void onStartClick(UseOnContext context, CompoundTag compoundTag) {
+		compoundTag.putString(TAG_TRANSPORT_MODE, ((BlockNode) context.getLevel().getBlockState(context.getClickedPos()).getBlock()).transportMode.toString());
+	}
+
+	@Override
+	protected void onEndClick(UseOnContext context, BlockPos posEnd, CompoundTag compoundTag) {
+		final Level world = context.getLevel();
+		final RailwayData railwayData = RailwayData.getInstance(world);
+		final BlockPos posStart = context.getClickedPos();
+		final BlockState stateStart = world.getBlockState(posStart);
+		final Block blockStart = stateStart.getBlock();
+		final BlockState stateEnd = world.getBlockState(posEnd);
+
+		if (railwayData != null && stateEnd.getBlock() instanceof BlockNode && ((BlockNode) blockStart).transportMode.toString().equals(compoundTag.getString(TAG_TRANSPORT_MODE))) {
+			final Player player = context.getPlayer();
+
+			if (isConnector) {
+				if (!posStart.equals(posEnd)) {
+					final float angle1 = BlockNode.getAngle(stateStart);
+					final float angle2 = BlockNode.getAngle(stateEnd);
+
+					final float angleDifference = (float) Math.toDegrees(Math.atan2(posEnd.getZ() - posStart.getZ(), posEnd.getX() - posStart.getX()));
+					final RailAngle railAngleStart = RailAngle.fromAngle(angle1 + (RailAngle.similarFacing(angleDifference, angle1) ? 0 : 180));
+					final RailAngle railAngleEnd = RailAngle.fromAngle(angle2 + (RailAngle.similarFacing(angleDifference, angle2) ? 180 : 0));
+
+					onConnect(world, context.getItemInHand(), ((BlockNode) blockStart).transportMode, stateStart, stateEnd, posStart, posEnd, railAngleStart, railAngleEnd, player, railwayData);
+				}
+			} else {
+				onRemove(world, posStart, posEnd, railwayData);
+			}
+		}
+
+		compoundTag.remove(TAG_TRANSPORT_MODE);
+	}
+
+	@Override
+	protected boolean clickCondition(UseOnContext context) {
+		final Level world = context.getLevel();
+		final Block blockStart = world.getBlockState(context.getClickedPos()).getBlock();
+		return blockStart instanceof BlockNode && (((BlockNode) blockStart).transportMode.continuousMovement && forContinuousMovementNode || !((BlockNode) blockStart).transportMode.continuousMovement && forNonContinuousMovementNode);
 	}
 
 	protected abstract void onConnect(Level world, ItemStack stack, TransportMode transportMode, BlockState stateStart, BlockState stateEnd, BlockPos posStart, BlockPos posEnd, RailAngle facingStart, RailAngle facingEnd, Player player, RailwayData railwayData);
